Changes method of populating asteroids
Instead of adding a set amount of asteroids, the game adds asteroids
until the sum of their values exceeds a set limit. At the start of the
game there is no difference, but later on the same method can be
called again to 'top off' the playing field with new asteroids, and
the limit can be gradually increased.