Fixed (LIKE FOR SERIOUS THIS TIME) collision.
Ship always calls the isCollidedWith method because it's the one which
will always have the mercy timer on it. It can't just check the other
object's mercy timer because non moving objects (pickups) don't have
them and it'll break the condition.

May want to fix that by making Pickup inherit MovingObject, but later.

Also moves actions relating to the ship dying to Ship.prototype.die()